package_checker/requirements.txt
click==7.1.2
termcolor==1.1.0
python-dateutil==2.8.2
GitPython==3.1.2
